some stuff

+++ database/migrations/2022_01_12_083821_create_photos_table.php
@@ -16,7 +16,7 @@ class CreatePhotosTable extends Migration
         Schema::create('photos', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('property_id')->unsigned()->index()->nullable();
+            $table->bigInteger('property_id')->unsigned()->index()->nullable();
             $table->foreign('property_id')->references('id')->on('property')->nullable()->unsigned()->index();
             $table->longText('photo_paths')->nullable();
         });
